mirror: make mirror legs on different tag pvs
The mirror leg could be placed on different PVs(with
different tags) from the other mirror legs with this patch.

We can't control the fist allocation when creating a new
mirrored lv now, because didn't restrict the areas into
parallel space.

This patch add the limit when add areas into parallel space,
so that mirror legs will always be placed on pvs with
differnt tags when creating a mirrored logical volume.

Signed-off-by: Guangliang Zhao <gzhao@suse.com>